requirements.txt
streamlit==1.30.0
networkx==3.2.1
tiktoken==0.5.2
OpenAI == 0.28.0
faiss-cpu==1.7.4
langchain==0.1.1
langchain-community==0.0.13
python-dotenv==1.0.0
beautifulsoup4==4.12.2
playwright==1.41.1
langchain-openai==0.0.5
sentence-transformers==2.3.1